Solve day 23 more efficently
This is done by condensing the graph, and the code now runs in less
than a minute.

Additionally, fix UB in the day 5 solution, and remove extra print
statements everywhere. Also upgrade Iros to fix a memory leak and
UB in TransparentString::insert(). The aoc solutions now pass ub and
address sanitizer.