Solve day 24
This was quite rough. For part 2 I resorted to using
running a python program which effectively runs sympy
to solve the problem. This appears to be the most common
approach, although it is possible to solve this problem
using a guassian elimination. This requires cancelling
out the collision time for each variable, and then
cancelling out the non-linear part (x*vx) by solving
for 2 lines at once.

      -------Part 1--------   --------Part 2--------
Day       Time  Rank  Score       Time   Rank  Score
 24   00:40:14   871      0   03:53:52   1784      0