Add script for downloading test input
The key insight is that all example input can be selected with the
following query:

.day-desc pre code

Of course, this finds all top-level blocks in the webpage and not just
the actual test input, so the user is forced to which block they want.
Most of the time this is block 0 but its not guaranteed. The advantage
of doing all blocks is it makes testing test inputs in part 2 easier,
since those also get downloaded.

Actually identifying the first test input possible using actual
web scraping library, but requires looking for test like "example"
in the previous paragraph. That isn't possible with shell scripting
tools I'm aware of.