Solve 2025 day 9
Spent 30 minutes remembering that when using the ray method for a
point in polygon test, you have to use on inclusive bound and one
exclusive bound or else it doesn't work properly when there are
horizontal lines...

Do this!

```cpp
if (y > ymin && y <= ymax) {
    intersections++;
}
```

Or!

```cpp
if (y >= ymin && y < ymax) {
    intersections++;
}
```

Not this!

```cpp
if (y >= ymin && y <= ymax) {
    intersections++;
}
```